register tokens and change parsing
